server: merge tftp.h into tftpd.c
